+++ tests/example_tests/example_parameter_dependencies_tests.py
@@ -1,0 +1,48 @@
+#!/usr/bin/env python3
+# JUBE Benchmarking Environment
+# Copyright (C) 2008-2022
+# Forschungszentrum Juelich GmbH, Juelich Supercomputing Centre
+# http://www.fz-juelich.de/jsc/jube
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+"""Test the cycle example"""
+
+from __future__ import (print_function,
+                        unicode_literals,
+                        division)
+
+import unittest
+from examples_tests import TestCase
+
+class TestParameterDependenciesExample(TestCase.TestExample):
+
+    """Class for testing the parameter_dependencies example"""
+
+    @classmethod
+    def setUpClass(cls):
+        '''
+        Automatically called method before tests in the class are run.
+
+        Create the necessary variables and paths for the specific example
+        '''
+        cls._name = "parameter_dependencies"
+        cls._stdout = [["hello 3 10", "hello 5 10", "world 1 20",
+                        "world 2 20", "world 4 20"],
+                         ["hello 3 10", "hello 5 10", "world 1 20",
+                        "world 2 20", "world 4 20"]]
+        super(TestParameterDependenciesExample, cls).setUpClass()
+        super(TestParameterDependenciesExample, cls)._execute_commands()
+
+if __name__ == "__main__":
+    unittest.main()